requirements.txt
prodict
pymongoext
pymongo
coinpaprika
redis
toolz
coinmarketcapapi
nh-currency
python-dotenv
pony
pychalk
aiohttp
mysql
requests
discord.py
psutil
asyncio
timeago
yarl<1.2
pytz
